Restructure, add missing file
