+++ metadata-transformer/src/test/java/no/unit/nva/metadata/service/testdata/MetaTagPair.java
@@ -1,0 +1,19 @@
+package no.unit.nva.metadata.service.testdata;
+
+public class MetaTagPair {
+    private final String name;
+    private final String content;
+
+    public MetaTagPair(String name, String value) {
+        this.name = name;
+        this.content = value;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public String getContent() {
+        return content;
+    }
+}
